lunchtime.media.mit.edu/lunchtime/requirements/apps.txt
Django
PIL
sorl-thumbnail==3.2.5
-e git+https://github.com/sciyoshi/pyfacebook.git#egg=facebook
httplib2
south
